cleanup/promtree/pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "promtree"
version = "0.1.4"
description = "PDF를 Markdown으로 변환하여 저장하는 라이브러리"

readme = "README.md"

authors = [
    { name = "Team.PromTree", email = "effort-result@naver.com" }
]

classifiers = [
    "Programming Language :: Python :: 3.12",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]

dependencies = [
    "PyMuPDF == 1.26.5",
    "surya-ocr == 0.17.0",
    "torch == 2.9.0",
    "transformers == 4.57.1",
    "numpy == 2.3.4",
    "Pillow == 10.4.0",
    "opencv-python-headless == 4.11.0.86",
    "einops == 0.8.1",
    "huggingface-hub == 0.36.0",
]

requires-python = "==3.12.10"

[project.urls]
Homepage = "https://github.com/Team-PromTree/promtree"
Repository = "https://github.io/Team-PromTree/promtree"

[tool.setuptools]
packages = { find = {} }

[project.scripts]
promtree = "promtree.__main__:main"
